Add model configurations and loop over these
There is a new flag to specify a model configurations file, wherein
multiple values for each model option can be specified. The main script
reads this file if given and sets up corresponding models to loop over.
Otherwise, it uses the arguments given on the command line or the
default values.