Add local 10x subset data set
Added entry in data set dictionary for 10x subset data set as well as
implemented function to load it.

Since there is local public URL for the 10x subset data set, the
data set download function was modified to support only an extension
as URL for now.

Support for local data set files should be improved at a later point.