Remove old model classes
Some model classes have not been updated along with the others to work
with the rest of implementation, since they did not work well with the
data. Instead of tediously updating them, they have been removed.

The two remaining models can then be more generalised to allow for hooks
for other models.