Fix overwriting evaluation log results
When a prediction method was used to predict cluster IDs and labels,
the training set was evaluated to get the latent values. But in doing
this, the evaluation log results were overwritten.

An argument for whether to log results has been added to the evaluate
methods of the model classes, and this is set to false in the above
case.