Disable standard deviations in model classes
Since the standard deviations are only used for a very small subset of
the evaluation set to plot profile comparisons, there is no need to
compute and use up memory to store them for all values.

The computation and storage of these have just been commented out for
now, since they are not essential.