Move computation of comparisons to analysis setup
The comparison arrays in the results analysis function are use by
multiple analyses and were previously only computed in the first one
(the metrics analysis). So if the latter analysis was excluded, the
script would crash. Therefore, the computation of the comparison arrays
have been moved the setup of the results analyses function at the start.