Plot histograms for sparse matrices
To use the histogram plotting functions, the value array or matrix has
to be flattened to a series. But if the values are represented as a
sparse matrix to conserve memory, converting to a dense series defeats
the purpose. Instead a count for the excess zeroes can provided. This is
added to the zero bin.

Also, a bug where the bars were not plotted when the *y* scale was
logarithmic has been fixed.