Add model evaluate argument to select output sets
The evaluate method for the model classes now takes a new argument
`output_versions` (default: "all"), which determines the sets
("transformed", "reconstructed", "latent") it returns.

`output_versions` should be a string or a list of strings. If it is a
list, the corresponding sets are returned in the same order, and if it
is a string only the corresponding set is returned (i.e. not wrapped in
a list). The only exception to this if the argument is "all", which
returns all sets in the default order as listed above.