Add linear factor models as alternative to MLP's
For the architectures of the inference and the generative models,
linear factor models (LFM have been added as an alternative to the
neural networks (MLP) normally used in a VAE.

Choosing a LFM architecture for either model parts of the VAE,
essentially just skips the hidden layers of MLP part of a standard VAE
for that model part.

To distinguish models using LFM's from models using MLP's, identifiers
were added to the model names, and these are in turn use in the cross-
analysis.

Lastly, two command-line arguments have been added to set the
architectures of the inference and the generative models:
`--inference-architecture` and `--generative-architecture`,
respectively. They both default to "MLP", and right now only that as
well as "LFM" are supported values.