Implement recovery from earlier temporary model
If the script crashed before during training and a temporary log
directory was used, the temporary log directory would not be copied to
the permanent log directory. Restarting the script would then recover
from the permanent log directory, ignoring the saved model parameters in
the temporary log directory.

Now, the script will recover from the temporary log directory, if the
model there have been trained for more epochs than the one in the
permanent log directory.