Make validation set optional when training
There was several places throughout the implementation where it was
assumed that three subsets (training, validation, and test set) as well
as computations related to each subset would be present. To compare
our model to other models which train and evaluate on the full data set,
it was necessary to make the validation set optional.

This means that intermediate plots when training are evaluated on the
training set and that early stopping is disabled when no validation set
is given.