Enable decomposition before prediction
Added an option to choose a different training set for the prediction
method. One can choose among the existing subsets ("training",
"validation", of "test") or the full data set ("full") by using command
argument "--prediction-training-set-name".

An option to decompose the latent representations of this prediction
training set and the evaluation set has also been added. Since this is
mainly for t-SNE decomposition, the latent representations have to be
fitted and transformed together. The method can be set using the command
argument "--prediction-decomposition-method", and the resulting
dimensionality can be set by
"--prediction-decomposition-dimensionality".