Add DBSCAN as a prediction method
DBSCAN has been added as a method to cluster values and predict labels.
Since it is transductive, it only trains and is evaluated on the
evaluation set, and since it is a density-based method, it ignores the
fixed number of classes.

To estimate the one of the parameters of DBSCAN, the Python module
kneedle used for knee-point detection was required.