Move input parsing functions to separate module
Function to parse input for data sets and retrieve or create
corresponding metadata have been moved to the data.parsing` module.
The `data_sets` dictionary with pre-populated metadata have been
converted to a JSON file.

Because keys in JSON files cannot be integers, and since labels are used
as keys in the JSON file, loaders for data sets with integer labels have
been changed to use string labels.