Enable specifying data format
Data format, title, and specifications of data set was added to the
`DataSet` class initialiser. Data format was added to specify a format
for the data if not present in the metadata, while the remaining
arguments was added to easier pass this information to subsets and
related sets.

Loader names were changed to conform with the string normaliser, and
handling of the data format were changed in parsing and loading
functions.

A command-line argument was also added to specify the data format:
`--format` or `-f`.