Move analysis functions into separate modules
A new subpackage, `analyses`, have been made, and analysis functions
have been moved to appropriate modules within.

The function `indices_for_evaluation_subset`, which is used to determine
a subset of examples for which perform expensive computations for
evaluation, have been moved to the `data.auxiliary` module.

Other modules have been updated to reflect these changes and better
avoid import loops. On that last note, an argument to the the train
method of the VAE models have been added to specify a function to
analyse intermediate results, and the explicit import of such a function
have been removed.