Change plots for correlations in the latent space
For high-dimensional latent spaces it is not feasible to plot all
correlations, so the way correlations in the latent space are plotted
has been changed. Only the most correlated latent pairs are plotted for
the high-dimensional latent spaces, while all correlations for smaller
ones are still plotted. In addition, the correlation matrix is also
plotted.