Add support for sparse matrices in HDF5 format
A separate loader function for sparse matrices in HDF5 format has been
added. It is modelled on the function to read data sets from 10x
Genomics and as such it expects a single directory containing arrays for
the sparse matrix (with names as for SciPy's CSR/CSC sparse matrix
format: "data", "indices", "indptr", "shape").

The function will also load example/cell and feature/gene names, if
found, as well as an optional labels in a separate TSV file.

This fixes issue #3.